+++ sdk/keyvault/azkeys/CHANGELOG.md
@@ -13,6 +13,7 @@
 * Methods `BeginDeleteKey` and `BeginRecoverDeletedKey` now return a `*runtime.Poller[T]` with their respective response types.
 * Option types with a `ResumeToken` field now take the token by value.
 * Renamed `CreateECKeyOptions.CurveName` to `.Curve`
+* Renamed `ReleaseKeyOptions.Enc` to `.Algorithm`
 
 ### Bugs Fixed